+++ infra/settings_app_infra.py
@@ -11,34 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-
-import datarobot as dr
+from datarobot_pulumi_utils.schema.apps import ApplicationSourceArgs
+from datarobot_pulumi_utils.schema.exec_envs import RuntimeEnvironments
 
 from docsassist.i18n import LanguageCode, LocaleSettings
-from infra.common.globals import GlobalRuntimeEnvironment
-from infra.common.schema import ApplicationSourceArgs
 from infra.settings_main import PROJECT_ROOT, project_name
 
-
-def get_app_template_id(name: str) -> str:
-    client = dr.client.get_client()
-    try:
-        templates = client.get(
-            "customTemplates/", params={"templateType": "customApplicationTemplate"}
-        ).json()
-        template_id: str = next(
-            template["id"] for template in templates["data"] if template["name"] == name
-        )
-        return template_id
-    except Exception as e:
-        raise ValueError(
-            f"Could not find the Application Template ID for {name}"
-        ) from e
-
-
 _application_path = PROJECT_ROOT / "frontend"
-
 
 source_files = [
     (str(f), str(f.relative_to(_application_path)))
@@ -76,7 +55,7 @@ if application_locale != LanguageCode.EN:
 
 app_source_args = ApplicationSourceArgs(
     resource_name=f"Guarded RAG App Source [{project_name}]",
-    base_environment_id=GlobalRuntimeEnvironment.PYTHON_39_STREAMLIT.value.id,
+    base_environment_id=RuntimeEnvironments.PYTHON_312_APPLICATION_BASE.value.id,
     files=source_files,
 ).model_dump(mode="json", exclude_none=True)