+++ utils/credentials.py
@@ -20,6 +20,7 @@ from typing import Any
 import pulumi
 import pulumi_datarobot as datarobot
 import pydantic
+from datarobot_pulumi_utils.schema.llms import LLMConfig, LLMs
 
 from docsassist.credentials import (
     AWSBedrockCredentials,
@@ -27,11 +28,7 @@ from docsassist.credentials import (
     DRCredentials,
     GoogleCredentials,
 )
-from infra.common.globals import GlobalLLM, LLMConfig
-
-from ..settings_main import project_name
-
-# from .aws_credential import AWSCredential
+from infra.settings_main import project_name
 
 
 def get_credential_runtime_parameter_values(
@@ -155,7 +152,7 @@ def get_credentials(
 ) -> DRCredentials | None:
     try:
         credentials: DRCredentials
-        if llm == GlobalLLM.DEPLOYED_LLM:
+        if llm == LLMs.DEPLOYED_LLM:
             return None
         if llm.credential_type == "azure":
             credentials = AzureOpenAICredentials()
@@ -164,13 +161,13 @@ def get_credentials(
                     import openai
 
                     lookup = {
-                        GlobalLLM.AZURE_OPENAI_GPT_3_5_TURBO.name: "gpt-35-turbo",
-                        GlobalLLM.AZURE_OPENAI_GPT_3_5_TURBO_16K.name: "gpt-35-turbo-16k",
-                        GlobalLLM.AZURE_OPENAI_GPT_4.name: "gpt-4",
-                        GlobalLLM.AZURE_OPENAI_GPT_4_32K.name: "gpt-4-32k",
-                        GlobalLLM.AZURE_OPENAI_GPT_4_O.name: "gpt-4o",
-                        GlobalLLM.AZURE_OPENAI_GPT_4_O_MINI.name: "gpt-4o-mini",
-                        GlobalLLM.AZURE_OPENAI_GPT_4_TURBO.name: "gpt-4-turbo",
+                        LLMs.AZURE_OPENAI_GPT_3_5_TURBO.name: "gpt-35-turbo",
+                        LLMs.AZURE_OPENAI_GPT_3_5_TURBO_16K.name: "gpt-35-turbo-16k",
+                        LLMs.AZURE_OPENAI_GPT_4.name: "gpt-4",
+                        LLMs.AZURE_OPENAI_GPT_4_32K.name: "gpt-4-32k",
+                        LLMs.AZURE_OPENAI_GPT_4_O.name: "gpt-4o",
+                        LLMs.AZURE_OPENAI_GPT_4_O_MINI.name: "gpt-4o-mini",
+                        LLMs.AZURE_OPENAI_GPT_4_TURBO.name: "gpt-4-turbo",
                     }
                     if (
                         credentials.azure_deployment is not None
@@ -202,7 +199,7 @@ def get_credentials(
                             Unable to run a successful test completion against deployment '{credentials.azure_deployment or lookup[llm.name]}'
                             on '{credentials.azure_endpoint}' with API version '{credentials.api_version or "2023-05-15"}'
                             with provided Azure OpenAI credentials. Please validate your credentials.
-                            
+
                             Please validate your credentials or check {__file__} for details.
                             """)
                     ) from e
@@ -211,11 +208,11 @@ def get_credentials(
             credentials = AWSBedrockCredentials()
             if test_credentials:
                 lookup = {
-                    GlobalLLM.ANTHROPIC_CLAUDE_3_HAIKU.name: "anthropic.claude-3-haiku-20240307-v1:0",
-                    GlobalLLM.ANTHROPIC_CLAUDE_3_SONNET.name: "anthropic.claude-3-sonnet-20240229-v1:0",
-                    GlobalLLM.ANTHROPIC_CLAUDE_3_OPUS.name: "anthropic.claude-3-opus-20240229-v1:0",
-                    GlobalLLM.AMAZON_TITAN.name: "amazon.titan-text-express-v1",
-                    GlobalLLM.ANTHROPIC_CLAUDE_2.name: "anthropic.claude-v2:1",
+                    LLMs.ANTHROPIC_CLAUDE_3_HAIKU.name: "anthropic.claude-3-haiku-20240307-v1:0",
+                    LLMs.ANTHROPIC_CLAUDE_3_SONNET.name: "anthropic.claude-3-sonnet-20240229-v1:0",
+                    LLMs.ANTHROPIC_CLAUDE_3_OPUS.name: "anthropic.claude-3-opus-20240229-v1:0",
+                    LLMs.AMAZON_TITAN.name: "amazon.titan-text-express-v1",
+                    LLMs.ANTHROPIC_CLAUDE_2.name: "anthropic.claude-v2:1",
                 }
                 if credentials.region_name is None:
                     pulumi.warn("AWS region not set. Using default 'us-west-1'.")
@@ -249,10 +246,9 @@ def get_credentials(
                 except Exception as e:
                     raise ValueError(
                         textwrap.dedent(f"""
-                            Unable to run a successful test completion against model '{lookup[llm.name]}' in region '{credentials.region_name or "us-west-1"}' 
+                            Unable to run a successful test completion against model '{lookup[llm.name]}' in region '{credentials.region_name or "us-west-1"}'
                             using request body '{request_body}' with provided AWS credentials.
-                            
-                            
+
                             Please validate your credentials or check {__file__} for details.
                             """)
                     ) from e
@@ -260,9 +256,9 @@ def get_credentials(
             credentials = GoogleCredentials()
             if test_credentials:
                 lookup = {
-                    GlobalLLM.GOOGLE_1_5_PRO.name: "gemini-1.5-pro-001",
-                    GlobalLLM.GOOGLE_BISON.name: "chat-bison@002",
-                    GlobalLLM.GOOGLE_GEMINI_1_5_FLASH.name: "gemini-1.5-flash-001",
+                    LLMs.GOOGLE_1_5_PRO.name: "gemini-1.5-pro-001",
+                    LLMs.GOOGLE_BISON.name: "chat-bison@002",
+                    LLMs.GOOGLE_GEMINI_1_5_FLASH.name: "gemini-1.5-flash-001",
                 }
                 try:
                     import openai